Rearranging directories and files
Just making order between scripts that won't be used but might help to reference.
Had an idea to implement what I call dynamic beat referencing and added a future implementations folder so that each idea that comes but is too far ahead, won't occupy space in my mind and be there in case I want to further try to implement it.

--Changed-- Scr_GameEvents_Legacy will contain the script in which the beat text counter implementation won't work because I couldn't manage (and wouldn't manage) to understand how to apply font to message by script. Using TMPro package this problem was easily fixed.

--UpNext-- Referencing beats on the screen so that my collegue can use those references to make the level design