Preparing to switch to Player Movement
Player Movement Syncing with music is an integral part of the work and I feel like at this point I should start working on the player movement and collisions with the terrain and the platforms in order to understand how to approach the movement, the pace of the game and the interactions so that everything can feel and look smooth.

--Added-- Player can now start from a different position than the (0,0,0);

--Changed-- Some comments about the math used have been fixed because of inconsistency

--Changed-- Due to the offset at which the player starts, the beats referenced in the editor were shifted to match the starting position of the player

Note: work will shift onto the player movements and interactions. Nontheless due to my basic knowledge and experience with git, I won't use other branches because I lack the knowledge of working with branch merges